v5.13.0 Release

+++ src/Tgstation.Server.Host/Components/Byond/ByondManager.cs
@@ -11,7 +11,6 @@ using Microsoft.Extensions.Logging;
 
 using Tgstation.Server.Api.Models;
 using Tgstation.Server.Host.Components.Events;
-using Tgstation.Server.Host.Extensions;
 using Tgstation.Server.Host.IO;
 using Tgstation.Server.Host.Jobs;
 using Tgstation.Server.Host.Utils;
@@ -105,8 +104,7 @@ namespace Tgstation.Server.Host.Components.Byond
 		/// <param name="version">The <see cref="Version"/> to validate.</param>
 		static void CheckVersionParameter(Version version)
 		{
-			if (version == null)
-				throw new ArgumentNullException(nameof(version));
+			ArgumentNullException.ThrowIfNull(version);
 
 			if (version.Build == 0)
 				throw new ArgumentException("version.Build cannot be 0!", nameof(version));
@@ -166,6 +164,7 @@ namespace Tgstation.Server.Host.Components.Byond
 						ActiveVersion?.ToString(),
 						stringVersion,
 					},
+					false,
 					cancellationToken);
 
 				ActiveVersion = version;
@@ -207,8 +206,7 @@ namespace Tgstation.Server.Host.Components.Byond
 		/// <inheritdoc />
 		public async Task DeleteVersion(JobProgressReporter progressReporter, Version version, CancellationToken cancellationToken)
 		{
-			if (progressReporter == null)
-				throw new ArgumentNullException(nameof(progressReporter));
+			ArgumentNullException.ThrowIfNull(progressReporter);
 
 			CheckVersionParameter(version);
 
@@ -236,7 +234,7 @@ namespace Tgstation.Server.Host.Components.Byond
 				await Task.WhenAny(
 					containerTask,
 					activeVersionUpdate)
-					.WithToken(cancellationToken);
+					.WaitAsync(cancellationToken);
 
 				if (containerTask.IsCompleted)
 					logger.LogTrace("All BYOND locks for {version} are gone", version);
@@ -303,26 +301,28 @@ namespace Tgstation.Server.Host.Components.Byond
 
 			var activeVersionBytesTask = GetActiveVersion();
 
-			using (await SemaphoreSlimContext.Lock(UserFilesSemaphore, cancellationToken))
-			{
-				// Create local cfg directory in case it doesn't exist
-				var localCfgDirectory = ioManager.ConcatPath(
-						byondInstaller.PathToUserByondFolder,
-						CfgDirectoryName);
-				await ioManager.CreateDirectory(
-					localCfgDirectory,
-					cancellationToken);
-
-				// Delete trusted.txt so it doesn't grow too large
-				var trustedFilePath =
-					ioManager.ConcatPath(
+			var byondDir = byondInstaller.PathToUserByondFolder;
+			if (byondDir != null)
+				using (await SemaphoreSlimContext.Lock(UserFilesSemaphore, cancellationToken))
+				{
+					// Create local cfg directory in case it doesn't exist
+					var localCfgDirectory = ioManager.ConcatPath(
+							byondDir,
+							CfgDirectoryName);
+					await ioManager.CreateDirectory(
 						localCfgDirectory,
-						TrustedDmbFileName);
-				logger.LogTrace("Deleting trusted .dmbs file {trustedFilePath}", trustedFilePath);
-				await ioManager.DeleteFile(
-					trustedFilePath,
-					cancellationToken);
-			}
+						cancellationToken);
+
+					// Delete trusted.txt so it doesn't grow too large
+					var trustedFilePath =
+						ioManager.ConcatPath(
+							localCfgDirectory,
+							TrustedDmbFileName);
+					logger.LogTrace("Deleting trusted .dmbs file {trustedFilePath}", trustedFilePath);
+					await ioManager.DeleteFile(
+						trustedFilePath,
+						cancellationToken);
+				}
 
 			await ioManager.CreateDirectory(DefaultIOManager.CurrentDirectory, cancellationToken);
 			var directories = await ioManager.GetDirectories(DefaultIOManager.CurrentDirectory, cancellationToken);
@@ -454,7 +454,7 @@ namespace Tgstation.Server.Host.Components.Byond
 					if (neededForLock && !installation.InstallationTask.IsCompleted)
 						logger.LogWarning("The required BYOND version ({version}) is not readily available! We will have to wait for it to install.", version);
 
-					await installation.InstallationTask.WithToken(cancellationToken);
+					await installation.InstallationTask.WaitAsync(cancellationToken);
 					return installLock;
 				}
 
@@ -477,7 +477,7 @@ namespace Tgstation.Server.Host.Components.Byond
 						progressReporter.StageName = "Running event";
 
 					var versionString = version.ToString();
-					await eventConsumer.HandleEvent(EventType.ByondInstallStart, new List<string> { versionString }, cancellationToken);
+					await eventConsumer.HandleEvent(EventType.ByondInstallStart, new List<string> { versionString }, false, cancellationToken);
 
 					await InstallVersionFiles(progressReporter, version, customVersionStream, cancellationToken);
 
@@ -486,7 +486,7 @@ namespace Tgstation.Server.Host.Components.Byond
 				catch (Exception ex)
 				{
 					if (ex is not OperationCanceledException)
-						await eventConsumer.HandleEvent(EventType.ByondInstallFail, new List<string> { ex.Message }, cancellationToken);
+						await eventConsumer.HandleEvent(EventType.ByondInstallFail, new List<string> { ex.Message }, false, cancellationToken);
 
 					lock (installedVersions)
 						installedVersions.Remove(version);
@@ -535,7 +535,7 @@ namespace Tgstation.Server.Host.Components.Byond
 				else
 					versionZipStream = customVersionStream;
 
-				using (versionZipStream)
+				await using (versionZipStream)
 				{
 					if (progressReporter != null)
 						progressReporter.StageName = "Cleaning target directory";
@@ -594,12 +594,16 @@ namespace Tgstation.Server.Host.Components.Byond
 				ioManager.ResolvePath(
 					ioManager.ConcatPath(
 						binPathForVersion,
-						byondInstaller.GetDreamDaemonName(version, out var supportsCli))),
+						byondInstaller.GetDreamDaemonName(
+							version,
+							out var supportsCli,
+							out var supportsMapThreads))),
 				ioManager.ResolvePath(
 					ioManager.ConcatPath(
 						binPathForVersion,
 						byondInstaller.DreamMakerName)),
-				supportsCli);
+				supportsCli,
+				supportsMapThreads);
 
 			var installationContainer = new ReferenceCountingContainer<ByondInstallation, ByondExecutableLock>(installation);
 
@@ -617,8 +621,15 @@ namespace Tgstation.Server.Host.Components.Byond
 		/// <returns>A <see cref="Task"/> representing the running operation.</returns>
 		async Task TrustDmbPath(string fullDmbPath, CancellationToken cancellationToken)
 		{
+			var byondDir = byondInstaller.PathToUserByondFolder;
+			if (String.IsNullOrWhiteSpace(byondDir))
+			{
+				logger.LogTrace("No relevant user BYOND directory to install a \"{fileName}\" in", TrustedDmbFileName);
+				return;
+			}
+
 			var trustedFilePath = ioManager.ConcatPath(
-				byondInstaller.PathToUserByondFolder,
+				byondDir,
 				CfgDirectoryName,
 				TrustedDmbFileName);
 
@@ -634,9 +645,7 @@ namespace Tgstation.Server.Host.Components.Byond
 					trustedFileText = $"{trustedFileText.Trim()}{Environment.NewLine}";
 				}
 				else
-				{
 					trustedFileText = String.Empty;
-				}
 
 				if (trustedFileText.Contains(fullDmbPath, StringComparison.Ordinal))
 					return;
